# openapi.yaml
openapi: 3.1.0
info:
  title: Qaida Places API
  version: 1.0.0
  summary: API inferred from the Qaida mobile UI kit screens.
  description: |
    This specification was derived from the HTML screens in the design folder.
    It describes the backend contract implied by the search, recommendations,
    place details, favorites, profile, collections, and notifications flows.

    Some parts are inferred from UI behavior and should be confirmed before implementation:
    - recommendation scoring logic
    - booking flows mentioned in notifications
    - collection vs default favorites behavior
    - push notification delivery strategy
servers:
  - url: https://api.qaida.app/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /user:
    get:
      tags: [User]
      summary: Get current user profile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [User]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/stats:
    get:
      tags: [User]
      summary: Get current user statistics
      responses:
        '200':
          description: User stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/favorites:
    get:
      tags: [Favorites]
      summary: List favorite places
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Favorite places
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/favorites/{placeId}:
    get:
      tags: [Favorites]
      summary: Check if place is in favorites
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      responses:
        '200':
          description: Favorite state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Favorites]
      summary: Add place to favorites
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      responses:
        '201':
          description: Place added to favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Favorites]
      summary: Remove place from favorites
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      responses:
        '200':
          description: Place removed from favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/collections:
    get:
      tags: [Collections]
      summary: List user collections
      responses:
        '200':
          description: Collections list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Collection'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Collections]
      summary: Create collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
      responses:
        '201':
          description: Collection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/collections/{collectionId}:
    get:
      tags: [Collections]
      summary: Get collection details
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '200':
          description: Collection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Collections]
      summary: Delete collection
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '204':
          description: Collection deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/collections/{collectionId}/places/{placeId}:
    post:
      tags: [Collections]
      summary: Add place to collection
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - $ref: '#/components/parameters/PlaceId'
      responses:
        '201':
          description: Place added to collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Collections]
      summary: Remove place from collection
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - $ref: '#/components/parameters/PlaceId'
      responses:
        '200':
          description: Place removed from collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/search-history:
    get:
      tags: [User]
      summary: List search history
      responses:
        '200':
          description: Search history list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchHistoryEntry'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [User]
      summary: Save search history entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSearchHistoryRequest'
      responses:
        '201':
          description: Search history entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchHistoryEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/search-history/{entryId}:
    delete:
      tags: [User]
      summary: Delete search history entry
      parameters:
        - in: path
          name: entryId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Search history entry deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /user/reviews/{reviewId}:
    delete:
      tags: [Reviews]
      summary: Delete current user review
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Review deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PlaceId:
      in: path
      name: placeId
      required: true
      schema:
        type: string
        format: uuid
    CollectionId:
      in: path
      name: collectionId
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    BudgetRange:
      type: string
      enum: ['$', '$$', '$$$']
    CompanyType:
      type: string
      enum: [solo, couple, friends, family]
    PlaceStatus:
      type: string
      enum: [open, closed, coming_soon]
    NotificationType:
      type: string
      enum: [new_place, reminder, saved_place, recommendation, booking]
...
    PlaceSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
        shortDescription:
          type: string
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        reviewCount:
          type: integer
          minimum: 0
        address:
          type: string
        distanceKm:
          type: number
          format: float
        budget:
          $ref: '#/components/schemas/BudgetRange'
        status:
          $ref: '#/components/schemas/PlaceStatus'
        imageUrl:
          type: string
          format: uri
        tags:
          type: array
          items:
            type: string
        isFavorite:
          type: boolean
      required:
        [id, name, category, rating, reviewCount, address, budget, status, isFavorite]
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          format: uri
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        notificationPreferences:
          $ref: '#/components/schemas/NotificationPreferences'
      required: [id, name, email, preferences, notificationPreferences]
    UserPreferences:
      type: object
      properties:
        companyType:
          $ref: '#/components/schemas/CompanyType'
        favoriteCategories:
          type: array
          items:
            type: string
        budget:
          $ref: '#/components/schemas/BudgetRange'
      additionalProperties: false
    UserStats:
      type: object
      properties:
        placesVisited:
          type: integer
          minimum: 0
        reviewsCount:
          type: integer
          minimum: 0
        favoritesCount:
          type: integer
          minimum: 0
      required: [placesVisited, reviewsCount, favoritesCount]
    FavoriteStatus:
      type: object
      properties:
        placeId:
          type: string
          format: uuid
        isFavorite:
          type: boolean
      required: [placeId, isFavorite]
    Collection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        placeCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
      required: [id, name, placeCount, createdAt]
    CollectionDetail:
      allOf:
        - $ref: '#/components/schemas/Collection'
        - type: object
          properties:
            places:
              type: array
              items:
                $ref: '#/components/schemas/PlaceSummary'
          required: [places]
...
    NotificationPreferences:
      type: object
      properties:
        enabled:
          type: boolean
        enabledTypes:
          type: array
          items:
            $ref: '#/components/schemas/NotificationType'
      required: [enabled, enabledTypes]
    SearchHistoryEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        query:
          type: string
        filters:
          $ref: '#/components/schemas/SearchFilters'
        createdAt:
          type: string
          format: date-time
      required: [id, query, filters, createdAt]
    SearchFilters:
      type: object
      properties:
        categories:
          type: array
          items:
            type: string
        companyType:
          $ref: '#/components/schemas/CompanyType'
        budget:
          $ref: '#/components/schemas/BudgetRange'
        distanceKm:
          type: number
          format: float
        sortBy:
          type: string
          enum: [popularity, rating, distance, recent]
      additionalProperties: false
...
    CreateCollectionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
      required: [name]
    CreateSearchHistoryRequest:
      type: object
      properties:
        query:
          type: string
        filters:
          $ref: '#/components/schemas/SearchFilters'
      required: [query, filters]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          format: uri
        preferences:
          $ref: '#/components/schemas/UserPreferences'
...
    PlaceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PlaceSummary'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
        offset:
          type: integer
      required: [items, total, limit, offset]
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
...